[refactory] ShareMemberVO 패키지 이동
