+++ PayStory/src/main/java/com/AourZ/PayStory/controller/MemberController.java
@@ -29,7 +29,6 @@ import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 import org.springframework.web.servlet.support.RequestContextUtils;
 
 import com.AourZ.PayStory.model.FileUtils;
-import com.AourZ.PayStory.model.board.BoardVO;
 import com.AourZ.PayStory.model.member.LoginVO;
 import com.AourZ.PayStory.model.member.MemberVO;
 import com.AourZ.PayStory.service.IMemberService;
